Add rotational support for future and support facing up/down on 1.13+
The rotational support is fully there, but has issues in that the maps
do not get rotated for an unknown reason that has not been able to be
diagnosed. This feature may be supported in the future, however, for now
it is simply dormant.

The facing up and down directions is supported but only for the default
orientation. This is because of the rotational support not functioning
properly and therefore, up and down are limited to facing north