Make PlayerConnectionProxy public
Making PlayerConnectionProxy public allows external sources
to get the sendPacket method and use it. Without this, they
get an error because the class cannot be accessed.

Previously, the sendPacket method was overridden in the
PlayerConnectionProxy for just this purpose. This, then
completes that purpose since it still had this barrier