Fix potions being cast to maps and add safeguards
Throwing a potion somehow can trigger the pick block
from creative inventories and if the ID of the potion is
high enough and matches a map this causes an exception
will completely stall out a netty thread.

This fixes the issue as well as putting a safeguard to
always notify the thread even if an exception is thrown.