Add 1.17 MinecraftVersion and invisible frames
Invisible frames added along with a config.yml entry to
change the default setting. This allows blank space to
not be rendered in 1.16+ and only render the pixels without
the maps or item frames behind them.

Invisible images can only be viewed from the front and
are completely invisible from the rear.