Locate image by contract and token id
Syncing an image deletion by copying the image Object to another
server is not guaranteed to work as the images may have different
entity ids on each server, meaning the deletion won't be correctly
sent to the Minecraft client. Instead, locate the server's Object
for that image using the unique contract and token id.