Allow custom render/hide distances for images (#56)
This change adds the ability to configure the distances for
rendering and hiding images. We needed a way to make some images 
visible at a greater player distance, so I've added a way to 
configure the distances instead of the hardcoded ones.

Fairly simple change and has been in use on our own servers for
 a while, finally got around to submitting this back