Update to Minecraft 1.21.4 version to 2.5.3 (#112)
* Update version to 2.5.3 with 1.21.4 support
* Fix pick item with new packet

In this release Minecraft added a second packet for picking items from entities that seems to replace the other packet. They didn't remove the other packet, but seemed to have stopped using it when picking items in creative.

Using the new packet, it's much more difficult to identify the exact item. Therefore, every feature is restored except for picking the same item if you already have it in your inventory; instead you'll get duplicates as you continue to pick items from the same section of an image.

---------

Signed-off-by: Mark <mark.vogel1@icloud.com>